Reversed hashed Agg implementation.
Only AGG_PLAIN and AGG_HASHED are reversed. The part relating to putting tuples
to hashtable is practically the same with hashtable lookups inlined.

To iterate over the hashtable effectively, method foreach was added to
simplehash.h. As in SeqScan or Hashjoin, the goal is to have only one loop
iterating over the tuples and sending them. We can have only one 'foreach' type
per hashtable in current implementation, this obviously should be changed if
needed.