Reversed hashed Agg implemented.
Only hashed Agg is reversed. The part relating to putting tuples to hashtable
is practically the same with hashtable lookups inlined.

To iterate over the hashtable effectively, method foreach added to
simplehash.h. As in SeqScan or Hashjoin, the goal is to have only one loop
iterating over the tuples and sending them.